SW-322: common function
