+++ src/components/Dashboard/ProjectManagement/SprintManagement/SprintCard.tsx
@@ -23,6 +23,7 @@ import { ToUpperCase } from "@/src/utils/helpers"
 import { TaskFiltersType } from "../types/taskFilters.type"
 import TaskFilters from "../TaskFilter/TaskFilters"
 import { TaskType } from "../constants/projectManagment"
+import { filterTasks } from "../utils/helper"
 
 interface Props {
   sprint: SelectSprint
@@ -88,28 +89,11 @@ export default function SprintCardPage({
   ])
 
   useEffect(() => {
-    if (!sprint.id) return
+    if (!sprint.id || !tasks?.length) return
 
-    if (filters && tasks?.length > 0) {
-      const filtered = tasks.filter((t) => {
-        return (
-          t?.sprint_id === sprint.id &&
-          (!filters.priority?.length ||
-            filters.priority.includes(t.task_priority)) &&
-          (!filters.type?.length || filters.type.includes(t.task_type)) &&
-          (!filters.status?.length ||
-            filters.status.includes(t.status_id || "")) &&
-          (!filters.assignee?.length ||
-            filters.assignee.includes(t.assign_to || "")) &&
-          (!filters.creator?.length ||
-            filters.creator.includes(t.created_by || ""))
-        )
-      })
+    const filtered = filterTasks(tasks, sprint.id, filters)
 
-      setFilteredTasks(filtered)
-    } else if (tasks?.length > 0) {
-      setFilteredTasks(tasks.filter((t) => t?.sprint_id === sprint.id))
-    }
+    setFilteredTasks(filtered)
   }, [tasks, sprint.id, filters])
 
   // PERMISSIONS INITATE
